requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
scikit-learn
pillow
numpy
python-multipart
